remove -y flag when calling soup
Never.  Ever.  Ever.  Ever cause a system to reboot without explicit consent from the administrator!  Ever!  That is very unfriendly behavior.

By calling soup with the -y flag here, it will not prompt before rebooting; the system will just disconnect.  If the state of the system is not such that it will survive a reboot, you've just caused an administrative headache.